+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## HEAD
+## 2.2.0 (2018-12-02)
 
 * Add `before_create`, `after_create`, `after_finish`, and `after_terminate` hooks (@janko-m)